+++ LRT/platform_desktop_linux/platform_data_queue.c
@@ -107,7 +107,7 @@ void platform_flushFIFO(UINT32 id){
 }
 
 
-void platform_writeFifo(UINT8 id, UINT32 addr, UINT32 size, UINT8* buffer) {
+void platform_writeFifo(UINT32 id, UINT32 addr, UINT32 size, UINT8* buffer) {
 	volatile UINT8* mutex = shMem_sync+id;
 
 	while(*mutex != 0);
@@ -115,7 +115,7 @@ void platform_writeFifo(UINT8 id, UINT32 addr, UINT32 size, UINT8* buffer) {
 }
 
 
-void platform_readFifo(UINT8 id, UINT32 addr, UINT32 size, UINT8* buffer) {
+void platform_readFifo(UINT32 id, UINT32 addr, UINT32 size, UINT8* buffer) {
 	volatile UINT8* mutex = shMem_sync+id;
 
 	while(*mutex != 1);
