+++ LRT/platform_include/platform_data_queue.h
@@ -40,7 +40,7 @@
 #include <platform_types.h>
 
 void platform_flushFIFO(UINT32 id);
-void platform_writeFifo(UINT8 id, UINT32 address, UINT32 size, UINT8* buffer);
-void platform_readFifo(UINT8 id, UINT32 address, UINT32 size, UINT8* buffer);
+void platform_writeFifo(UINT32 id, UINT32 address, UINT32 size, UINT8* buffer);
+void platform_readFifo(UINT32 id, UINT32 address, UINT32 size, UINT8* buffer);
 
 #endif /* HWQUEUES_H_ */
